R implementation complete.
I added the benchmark function for R implementation and also converted all the
named lists in other functions to named double vectors. The latter change cleans
up the code and makes it more similar to the Python implementation (which uses
Dictionaries). I did some cursory tests of the benchmark function, and it
appears to work.